+++ data/ben.txt
@@ -1,6 +1,1 @@
-T | 0 | 1
-T | 0 | 2
-T | 0 | 3
-T | 0 | 4
-D | 0 | 1 | 2
-E | 0 | 1 | 2 | 3
+D | 0 | return book | 2019-12-01
